+++ CHANGES.rst
@@ -4,6 +4,11 @@ Changelog
 2.0 (unreleased)
 ----------------
 
+- On the listing, respect the SchemaContext label if supplied. Otherwise,
+  build one based on the title. This gives us more label flexibility
+  outside the dexterity schemaeditor context.
+  [smcmahon]
+
 - Add fieldsWhichCannotBeDeleted attribute to SchemaContext, which can be
   used to disallow removal of certain fields.
   [davisagli]